+++ yandex/ig_structures_test.go
@@ -486,7 +486,7 @@ func TestFlattenInstanceGroupHealthChecks(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			res, err := flattenInstanceGroupHealthChecks(&instancegroup.InstanceGroup{HealthChecksSpec: tt.spec})
+			res, _, err := flattenInstanceGroupHealthChecks(&instancegroup.InstanceGroup{HealthChecksSpec: tt.spec})
 
 			if err != nil {
 				t.Errorf("%v", err)